+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   * new Cats avatars
   * new Birds avatars
   * new Fenecs avatars
+  * new Abstract avatars
   * plugin settings to change the avatar set (for now it is only possible to choose on the instance level, channel option will come later)
 
 ### Minor changes and fixes